Create LocalDialog component
