+++ ui/src/components/LocalDialog/index.tsx
@@ -1,41 +1,39 @@
-import { HTMLAttributes, createContext, useContext, useState } from "react";
+import * as DialogPrimitive from "@radix-ui/react-dialog";
 
+import { DialogProps } from "@radix-ui/react-dialog";
+import { PropsWithChildren } from "react";
 import { twMergeClsx } from "~/util/helpers";
 
-type LocalDialogContainerContextValue = {
-  container: HTMLDivElement | null;
-};
-
-const LocalDialogContainerContext =
-  createContext<LocalDialogContainerContextValue | null>(null);
-
-LocalDialogContainerContext.displayName = "LocalDialogContainerContext";
-
-export const useLocalDialogContainer = () => {
-  const context = useContext(LocalDialogContainerContext);
-  if (!context)
-    throw new Error(
-      "useLocalDialogContainer must be used inside <LocalDialogContainer>"
-    );
-  return context;
-};
-
-type LocalDialogContainerProps = HTMLAttributes<HTMLDivElement>;
-
-export const LocalDialogContainer = ({
+export const LocalDialog = ({
   children,
-  className,
-}: LocalDialogContainerProps) => {
-  const [container, setContainer] = useState<HTMLDivElement | null>(null);
+  onOpenChange,
+}: PropsWithChildren & Pick<DialogProps, "onOpenChange">) => (
+  <DialogPrimitive.Root modal={false} onOpenChange={onOpenChange}>
+    {children}
+  </DialogPrimitive.Root>
+);
 
-  return (
-    <LocalDialogContainerContext.Provider value={{ container }}>
-      <div
-        ref={setContainer}
-        className={twMergeClsx("relative isolate z-0", className)}
+export const LocalDialogContent = ({
+  children,
+  container,
+}: PropsWithChildren & { container: HTMLDivElement | null }) => (
+  <DialogPrimitive.DialogPortal container={container}>
+    <div
+      className="absolute inset-0 flex items-start justify-center px-5 pt-16"
+      onClick={(event) => event.stopPropagation()}
+    >
+      <div className="absolute inset-0 -m-2 bg-black/10 backdrop-blur-sm" />
+      <DialogPrimitive.Content
+        className={twMergeClsx(
+          "pointer-events-auto fixed z-50 grid w-full gap-4 rounded-b-lg bg-gray-1 p-6 animate-in data-[state=open]:fade-in-90 data-[state=open]:slide-in-from-bottom-10 sm:max-w-lg sm:rounded-lg sm:zoom-in-90 data-[state=open]:sm:slide-in-from-bottom-0",
+          "dark:bg-gray-dark-1"
+        )}
+        onInteractOutside={(event) => {
+          event.preventDefault();
+        }}
       >
         {children}
-      </div>
-    </LocalDialogContainerContext.Provider>
-  );
-};
+      </DialogPrimitive.Content>
+    </div>
+  </DialogPrimitive.DialogPortal>
+);
